fix: Fix bolus progress reporter, rework logging & sync delivery limits

+++ Dependencies/DanaKit/DanaKit/Packets/DanaBolusGetStepInformation.swift
@@ -8,10 +8,10 @@
 
 struct PacketBolusGetStepInformation {
     var bolusType: UInt8
-    var initialBolusAmount: Float
+    var initialBolusAmount: Double
     var lastBolusTime: Date
-    var lastBolusAmount: Float
-    var maxBolus: Float
+    var lastBolusAmount: Double
+    var maxBolus: Double
     var bolusStep: UInt8
 }
 
@@ -26,10 +26,10 @@ func parsePacketBolusGetStepInformation(data: Data) -> DanaParsePacket<PacketBol
 
     return DanaParsePacket(success: data[DataStart] == 0, rawData: data, data: PacketBolusGetStepInformation(
         bolusType: data[DataStart + 1],
-        initialBolusAmount: Float(data.uint16(at: DataStart + 2)) / 100,
+        initialBolusAmount: Double(data.uint16(at: DataStart + 2)) / 100,
         lastBolusTime: lastBolusTime,
-        lastBolusAmount: Float(data.uint16(at: DataStart + 6)) / 100,
-        maxBolus: Float(data.uint16(at: DataStart + 8)) / 100,
+        lastBolusAmount: Double(data.uint16(at: DataStart + 6)) / 100,
+        maxBolus: Double(data.uint16(at: DataStart + 8)) / 100,
         bolusStep: data[DataStart + 10]
     ))
 }
